samples-master/python/level1_single_api/9_amct/amct_pytorch/requirements.txt
numpy
Pillow
torch==1.10.0
onnx==1.9.0
onnxruntime==1.6.0
onnxruntime-gpu==1.8.0
opencv-python==4.9.0.80
protobuf==3.20.0
torchvision==0.11.1